+++ core/acp-packet/src/test/kotlin/pers/acp/test/TestXml.kt
@@ -1,0 +1,22 @@
+package pers.acp.test
+
+import org.junit.jupiter.api.Test
+import pers.acp.packet.xml.XmlPacket
+
+class TestXml {
+    @Test
+    fun testXmlToObj() {
+        val xmlStr = """
+            <?xml version="1.0" encoding="UTF-8"?>
+            <root>
+                <fileFlag>123</fileFlag>
+                <BODY>
+                    <mobile>33333333</mobile>
+                </BODY>
+            </root>
+        """.trimIndent()
+        val obj = XmlPacket.xmlToObject(xmlStr,TestBean::class.java)
+        println(obj)
+        println(XmlPacket.objectToXML(obj!!))
+    }
+}