+++ cloud/acp-spring-cloud-starter/src/main/kotlin/pers/acp/spring/cloud/log/LogInfo.kt
@@ -44,9 +44,6 @@ class LogInfo {
     var message: String? = null
         internal set
 
-    var throwable: Throwable? = null
-        internal set
-
     var params: List<Any?> = listOf()
         internal set